split table + styles
