+++ packages/syft/src/syft/types/syft_object.py
@@ -44,7 +44,7 @@ from ..serde.recursive_primitives import recursive_serde_register_type
 from ..serde.serialize import _serialize as serialize
 from ..util.autoreload import autoreload_enabled
 from ..util.markdown import as_markdown_python_code
-from ..util.notebook_ui.notebook_addons import create_table_template
+from ..util.notebook_ui.components.table import create_table_template
 from ..util.util import aggressive_set_attr
 from ..util.util import full_name_with_qualname
 from ..util.util import get_qualname_for
